# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a megrez backend server.",
        "title": "Megrez backend API",
        "termsOfService": "http://megrez.run",
        "contact": {
            "name": "Megrez",
            "url": "http://megrez.run",
            "email": "alkaidchen@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/comment": {
            "post": {
                "description": "create comment",
                "consumes": [
                    "application/json"
                ],
                "summary": "create comment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateCommentForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/errmsg.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/comment/{id}": {
            "delete": {
                "description": "delete comment by comment id",
                "consumes": [
                    "application/json"
                ],
                "summary": "delete comment by comment id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "comment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/errmsg.Response"
                        }
                    }
                }
            }
        },
        "/api/admin/comments": {
            "get": {
                "description": "list comments",
                "consumes": [
                    "application/json"
                ],
                "summary": "list comments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page num",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.Pagination"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.CommentListDTO"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/admin/install": {
            "post": {
                "description": "install blog form",
                "consumes": [
                    "application/json"
                ],
                "summary": "install blog form",
                "parameters": [
                    {
                        "description": "install blog form",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InstallBlogForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ArticleDTO": {
            "type": "object",
            "properties": {
                "allowedComment": {
                    "type": "boolean"
                },
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "cover": {
                    "type": "string"
                },
                "editTime": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isTop": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "publishTime": {
                    "type": "string"
                },
                "seoDescription": {
                    "type": "string"
                },
                "seoKeywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "title": {
                    "type": "string"
                },
                "visits": {
                    "type": "integer"
                },
                "wordCount": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CommentListDTO": {
            "type": "object",
            "properties": {
                "article": {
                    "$ref": "#/definitions/dto.ArticleDTO"
                },
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "mail": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/dto.PageDTO"
                },
                "parentID": {
                    "type": "integer"
                },
                "rootID": {
                    "type": "integer"
                },
                "site": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "dto.CreateCommentForm": {
            "type": "object",
            "properties": {
                "articleID": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "pageID": {
                    "type": "integer"
                },
                "parentID": {
                    "type": "integer"
                },
                "rootID": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        },
...
        "dto.InstallBlogForm": {
            "type": "object",
            "properties": {
                "blogTitle": {
                    "type": "string"
                },
                "blogURL": {
                    "type": "string"
                },
                "confirmPassword": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.PageDTO": {
            "type": "object",
            "properties": {
                "cover": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "formatContent": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "likes": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "originalContent": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "visits": {
                    "type": "integer"
                }
            }
        },
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "list": {},
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "errmsg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
    }
}